+++ ios/RNDeviceInfo/RNDeviceInfo.m
@@ -9,8 +9,10 @@
 #include <ifaddrs.h>
 #include <arpa/inet.h>
 #import <mach-o/arch.h>
+#import <React/RCTUtils.h>
 #import "RNDeviceInfo.h"
 #import "DeviceUID.h"
+
 #if !(TARGET_OS_TV)
 #import <LocalAuthentication/LocalAuthentication.h>
 #endif
@@ -24,25 +26,58 @@ typedef NS_ENUM(NSInteger, DeviceType) {
 
 #define DeviceTypeValues [NSArray arrayWithObjects: @"Handset", @"Tablet", @"Tv", @"Unknown", nil]
 
-@interface RNDeviceInfo()
-@property (nonatomic) bool isEmulator;
-@end
-
 #if !(TARGET_OS_TV)
 @import CoreTelephony;
+@import Darwin.sys.sysctl;
 #endif
 
 @implementation RNDeviceInfo
+{
+    bool hasListeners;
+}
 
 @synthesize isEmulator;
 
-RCT_EXPORT_MODULE(RNDeviceInfo)
+RCT_EXPORT_MODULE();
 
 + (BOOL)requiresMainQueueSetup
 {
    return YES;
 }
 
+- (NSArray<NSString *> *)supportedEvents
+{
+    return @[@"batteryLevelDidChange", @"batteryLevelIsLow", @"powerStateDidChange"];
+}
+
+- (id)init
+{
+    if ((self = [super init])) {
+#if !TARGET_OS_TV
+        _lowBatteryThreshold = 20;
+        [[UIDevice currentDevice] setBatteryMonitoringEnabled:YES];
+
+        [[NSNotificationCenter defaultCenter] addObserver:self
+                                                 selector:@selector(batteryLevelDidChange:)
+                                                     name:UIDeviceBatteryLevelDidChangeNotification
+                                                   object: nil];
+        [[NSNotificationCenter defaultCenter] addObserver:self
+                                                 selector:@selector(powerStateDidChange:)
+                                                     name:UIDeviceBatteryStateDidChangeNotification
+                                                   object: nil];
+#endif
+    }
+
+    return self;
+}
+
+- (void)startObserving {
+    hasListeners = YES;
+}
+
+- (void)stopObserving {
+    hasListeners = NO;
+}
 
 - (NSString*) deviceId
 {
@@ -144,6 +179,14 @@ RCT_EXPORT_MODULE(RNDeviceInfo)
                               @"iPad7,4"   :@"iPad Pro 10.5-inch",// iPad Pro 10.5-inch - Cellular
                               @"iPad7,5"   :@"iPad (6th generation)",// iPad (6th generation) - Wifi
                               @"iPad7,6"   :@"iPad (6th generation)",// iPad (6th generation) - Cellular
+                              @"iPad8,1"   :@"iPad Pro 11-inch (3rd generation)", // iPad Pro 11 inch (3rd generation) - Wifi
+                              @"iPad8,2"   :@"iPad Pro 11-inch (3rd generation)", // iPad Pro 11 inch (3rd generation) - 1TB - Wifi
+                              @"iPad8,3"   :@"iPad Pro 11-inch (3rd generation)", // iPad Pro 11 inch (3rd generation) - Wifi + cellular
+                              @"iPad8,4"   :@"iPad Pro 11-inch (3rd generation)", // iPad Pro 11 inch (3rd generation) - 1TB - Wifi + cellular
+                              @"iPad8,5"   :@"iPad Pro 12.9-inch (3rd generation)", // iPad Pro 12.9 inch (3rd generation) - Wifi
+                              @"iPad8,6"   :@"iPad Pro 12.9-inch (3rd generation)", // iPad Pro 12.9 inch (3rd generation) - 1TB - Wifi
+                              @"iPad8,7"   :@"iPad Pro 12.9-inch (3rd generation)", // iPad Pro 12.9 inch (3rd generation) - Wifi + cellular
+                              @"iPad8,8"   :@"iPad Pro 12.9-inch (3rd generation)", // iPad Pro 12.9 inch (3rd generation) - 1TB - Wifi + cellular
                               @"AppleTV2,1":@"Apple TV",        // Apple TV (2nd Generation)
                               @"AppleTV3,1":@"Apple TV",        // Apple TV (3rd Generation)
                               @"AppleTV3,2":@"Apple TV",        // Apple TV (3rd Generation - Rev A)
@@ -197,11 +240,16 @@ RCT_EXPORT_MODULE(RNDeviceInfo)
 
 - (NSString*) deviceLocale
 {
-    NSString *language = [[NSLocale preferredLanguages] objectAtIndex:0];
+    NSString *language = [[NSLocale preferredLanguages] firstObject];
     return language;
 }
 
-- (NSString*) deviceCountry
+- (NSArray<NSString *> *) preferredLocales
+{
+    return [NSLocale preferredLanguages];
+}
+
+- (NSString*) deviceCountry 
 {
   NSString *country = [[NSLocale currentLocale] objectForKey:NSLocaleCountryCode];
   return country;
@@ -232,20 +280,25 @@ RCT_EXPORT_MODULE(RNDeviceInfo)
 - (NSNumber*) fontScale
 {
   float fontScale = 1.0;
-  NSString *contentSize = [UIApplication sharedApplication].preferredContentSizeCategory;
+  UIApplication *application = RCTSharedApplication();
 
-  if ([contentSize isEqual: @"UICTContentSizeCategoryXS"]) fontScale = 0.82;
-  else if ([contentSize isEqual: @"UICTContentSizeCategoryS"]) fontScale = 0.88;
-  else if ([contentSize isEqual: @"UICTContentSizeCategoryM"]) fontScale = 0.95;
-  else if ([contentSize isEqual: @"UICTContentSizeCategoryL"]) fontScale = 1.0;
-  else if ([contentSize isEqual: @"UICTContentSizeCategoryXL"]) fontScale = 1.12;
-  else if ([contentSize isEqual: @"UICTContentSizeCategoryXXL"]) fontScale = 1.23;
-  else if ([contentSize isEqual: @"UICTContentSizeCategoryXXXL"]) fontScale = 1.35;
-  else if ([contentSize isEqual: @"UICTContentSizeCategoryAccessibilityM"]) fontScale = 1.64;
-  else if ([contentSize isEqual: @"UICTContentSizeCategoryAccessibilityL"]) fontScale = 1.95;
-  else if ([contentSize isEqual: @"UICTContentSizeCategoryAccessibilityXL"]) fontScale = 2.35;
-  else if ([contentSize isEqual: @"UICTContentSizeCategoryAccessibilityXXL"]) fontScale = 2.76;
-  else if ([contentSize isEqual: @"UICTContentSizeCategoryAccessibilityXXXL"]) fontScale = 3.12;
+  // Shared application is unavailable in an app extension.
+  if (application) {
+    NSString *contentSize = application.preferredContentSizeCategory;
+
+    if ([contentSize isEqual: @"UICTContentSizeCategoryXS"]) fontScale = 0.82;
+    else if ([contentSize isEqual: @"UICTContentSizeCategoryS"]) fontScale = 0.88;
+    else if ([contentSize isEqual: @"UICTContentSizeCategoryM"]) fontScale = 0.95;
+    else if ([contentSize isEqual: @"UICTContentSizeCategoryL"]) fontScale = 1.0;
+    else if ([contentSize isEqual: @"UICTContentSizeCategoryXL"]) fontScale = 1.12;
+    else if ([contentSize isEqual: @"UICTContentSizeCategoryXXL"]) fontScale = 1.23;
+    else if ([contentSize isEqual: @"UICTContentSizeCategoryXXXL"]) fontScale = 1.35;
+    else if ([contentSize isEqual: @"UICTContentSizeCategoryAccessibilityM"]) fontScale = 1.64;
+    else if ([contentSize isEqual: @"UICTContentSizeCategoryAccessibilityL"]) fontScale = 1.95;
+    else if ([contentSize isEqual: @"UICTContentSizeCategoryAccessibilityXL"]) fontScale = 2.35;
+    else if ([contentSize isEqual: @"UICTContentSizeCategoryAccessibilityXXL"]) fontScale = 2.76;
+    else if ([contentSize isEqual: @"UICTContentSizeCategoryAccessibilityXXXL"]) fontScale = 3.12;
+  }
 
   return [NSNumber numberWithFloat: fontScale];
 }
@@ -289,9 +342,24 @@ RCT_EXPORT_MODULE(RNDeviceInfo)
 
 - (NSString *)getCPUType {
     /* https://stackoverflow.com/questions/19859388/how-can-i-get-the-ios-device-cpu-architecture-in-runtime */
-    NXArchInfo *info = NXGetLocalArchInfo();
+    const NXArchInfo *info = NXGetLocalArchInfo();
     NSString *typeOfCpu = [NSString stringWithUTF8String:info->description];
     return typeOfCpu;
+}
+
+
+- (NSString *)getBuildId {
+    #if TARGET_OS_TV
+        return @"not available";
+    #else
+        size_t bufferSize = 64;
+        NSMutableData *buffer = [[NSMutableData alloc] initWithLength:bufferSize];
+        int status = sysctlbyname("kern.osversion", buffer.mutableBytes, &bufferSize, NULL, 0);
+        if (status != 0) {
+            return @"not available";
+        }
+        return [[NSString alloc] initWithCString:buffer.mutableBytes encoding:NSUTF8StringEncoding];
+    #endif
 }
 
 - (NSDictionary *)constantsToExport
@@ -307,12 +375,14 @@ RCT_EXPORT_MODULE(RNDeviceInfo)
              @"deviceId": self.deviceId ?: [NSNull null],
              @"deviceName": currentDevice.name,
              @"deviceLocale": self.deviceLocale ?: [NSNull null],
+             @"preferredLocales": self.preferredLocales ?: [NSNull null],
              @"deviceCountry": self.deviceCountry ?: [NSNull null],
              @"uniqueId": uniqueId,
              @"appName": [[NSBundle mainBundle] objectForInfoDictionaryKey:@"CFBundleDisplayName"] ?: [NSNull null],
              @"bundleId": [[NSBundle mainBundle] objectForInfoDictionaryKey:@"CFBundleIdentifier"] ?: [NSNull null],
              @"appVersion": [[NSBundle mainBundle] objectForInfoDictionaryKey:@"CFBundleShortVersionString"] ?: [NSNull null],
              @"buildNumber": [[NSBundle mainBundle] objectForInfoDictionaryKey:@"CFBundleVersion"] ?: [NSNull null],
+             @"buildId": [self getBuildId],
              @"systemManufacturer": @"Apple",
              @"carrier": self.carrier ?: [NSNull null],
              @"userAgent": self.userAgent ?: [NSNull null],
@@ -376,14 +446,74 @@ RCT_EXPORT_METHOD(isPinOrFingerprintSet:(RCTResponseSenderBlock)callback)
     callback(@[[NSNumber numberWithBool:isPinOrFingerprintSet]]);
 }
 
+- (void)batteryLevelDidChange:(NSNotification *)notification
+{
+    if (!hasListeners) {
+        return;
+    }
+
+    float batteryLevel = [self.powerState[@"batteryLevel"] floatValue];
+    [self sendEventWithName:@"batteryLevelDidChange" body:@(batteryLevel)];
+
+    if (batteryLevel <= _lowBatteryThreshold) {
+        [self sendEventWithName:@"batteryLevelIsLow" body:@(batteryLevel)];
+    }
+}
+
+- (void)powerStateDidChange:(NSNotification *)notification
+{
+    if (!hasListeners) {
+        return;
+    }
+
+    [self sendEventWithName:@"powerStateDidChange" body:self.powerState];
+}
+
+- (NSDictionary *)powerState
+{
+#if RCT_DEV && (!TARGET_IPHONE_SIMULATOR)
+    if ([UIDevice currentDevice].isBatteryMonitoringEnabled != true) {
+        RCTLogWarn(@"Battery monitoring is not enabled. "
+                   "You need to enable monitoring with `[UIDevice currentDevice].batteryMonitoringEnabled = TRUE`");
+    }
+#endif
+#if RCT_DEV && TARGET_IPHONE_SIMULATOR && !TARGET_OS_TV
+    if ([UIDevice currentDevice].batteryState == UIDeviceBatteryStateUnknown) {
+        RCTLogWarn(@"Battery state `unknown` and monitoring disabled, this is normal for simulators and tvOS.");
+    }
+#endif
+
+    return @{
+#if TARGET_OS_TV
+             @"batteryLevel": @1,
+             @"batteryState": @"full",
+#else
+             @"batteryLevel": @([UIDevice currentDevice].batteryLevel),
+             @"batteryState": [@[@"unknown", @"unplugged", @"charging", @"full"] objectAtIndex: [UIDevice currentDevice].batteryState],
+             @"lowPowerMode": @([NSProcessInfo processInfo].isLowPowerModeEnabled),
+#endif
+             };
+}
+
 RCT_EXPORT_METHOD(getBatteryLevel:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject)
 {
-  #if TARGET_OS_TV
-    float batteryLevel = 1.0;
-  #else
-    float batteryLevel = [UIDevice currentDevice].batteryLevel;
-  #endif
-    resolve(@(batteryLevel));
+    resolve(@([self.powerState[@"batteryLevel"] floatValue]));
+}
+
+RCT_EXPORT_METHOD(getPowerState:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject)
+{
+    return resolve(self.powerState);
+}
+
+RCT_EXPORT_METHOD(isBatteryCharging:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject)
+{
+    BOOL isCharging = [self.powerState[@"batteryState"] isEqualToString:@"charging"];
+    resolve(@(isCharging));
+}
+
+- (void)dealloc
+{
+    [[NSNotificationCenter defaultCenter] removeObserver:self];
 }
 
 @end